Add convenience method to parse X509 certificates
